# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Quizlet API (v1)
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            unauthorized:
              value: { code: "unauthorized", message: "Не авторизован" }
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            conflict:
              value: { code: "conflict", message: "Ресурс уже существует" }
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            validation_error:
              value:
                code: "validation_error"
                message: "Проверьте корректность введённых данных"
                details: { field: "ошибка" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    RegisterRequest:
      type: object
      required: [email, login, password]
      properties:
        email: { type: string, format: email }
        login:
          type: string
          minLength: 3
          maxLength: 64
          example: "login"
        password:
          type: string
          minLength: 8
          maxLength: 64
          pattern: "^[A-Za-z0-9_.-]+$"
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: { type: string }
        password: { type: string }
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      security: []  # публично
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Вход
      security: [] 
      requestBody:
        required: true
        content:
          application/json: 
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Выход
      parameters: []
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
...